restructure has utils
