+++ src/utils/hiveAuthenticationService/messages/authenticate.ts
@@ -3,13 +3,14 @@ import Crypto from 'crypto-js';
 import {ModalComponent} from 'utils/modal.enum';
 import {navigate} from 'utils/navigation';
 import HAS from '..';
+import {answerAuthReq} from '../helpers/auth';
 import {HAS_AuhtDecrypted, HAS_AuthPayload} from '../payloads.types';
 
 export const processAuthenticationRequest = (
   has: HAS,
   payload: HAS_AuthPayload,
 ) => {
-  has.checkPayload(payload);
+  HAS.checkPayload(payload);
   const accountSession = HAS.findSessionByUUID(payload.uuid);
   assert(accountSession, 'This account has not been connected through HAS.');
   console.log(accountSession, payload);
@@ -23,6 +24,6 @@ export const processAuthenticationRequest = (
 
   navigate('ModalScreen', {
     name: ModalComponent.HAS_AUTH,
-    data: {...payload, callback: has.answerAuthReq},
+    data: {...payload, has, callback: answerAuthReq},
   });
 };
